+++ src/components/func/FormFunc.module.sass
@@ -1,2 +1,3 @@
 .border
+    width: 25%
     border: 1px solid red
