--- src/index.css
@@ -1,1 +1,0 @@
-@use '~normalize.css';
