+++ src/index.sass
@@ -4,3 +4,12 @@
     color: purple
     margin: 50px auto
     font-size: 50px
+
+button
+    background-color: #ffb8b8
+    border: 2px solid #e9a0a0
+    margin: 2px 10px
+
+.DZ
+    margin: 40px 0 10px 0
+    color: purple
